+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added 
+
+- `opentelemetry-instrumentation-requests`: add ability to capture custom headers
+  ([#3987](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3987))
+
 ## Version 1.39.0/0.60b0 (2025-12-03)
 
 ### Added