+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `opentelemetry-instrumentation-asyncpg`: Hydrate span attributes before creation so samplers can filter on database details
+  ([#3643](https://github.com/open-telemetry/opentelemetry-python-contrib/issues/3643))
+
 ## Version 1.39.0/0.60b0 (2025-12-03)
 
 ### Added 
@@ -73,9 +78,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3936](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3936))
 - `opentelemetry-instrumentation-aiohttp-client`: Update instrumentor to respect suppressing http instrumentation
   ([#3957](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3957))
-- `opentelemetry-instrumentation-asyncpg`: Hydrate span attributes before creation so samplers can filter on database details
-  ([#3643](https://github.com/open-telemetry/opentelemetry-python-contrib/issues/3643))
-
 ## Version 1.38.0/0.59b0 (2025-10-16)
 
 ### Fixed